Make busybox configuration consistent
The busybox configuration provided by Asus was designed against an
earlier version of busybox. This created some inconsistencies, such as
leaving CONFIG_FEATURE_LS_COLOR unset while setting
CONFIG_FEATURE_LS_COLOR_IS_DEFAULT to be the default setting. This does
not cause a problem in the case of ls, but it is better for this file to
be consistent.

Signed-off-by: Richard Yao <ryao@cs.stonybrook.edu>