Force the restart of dnsmasq when restarting it to enable/disable web redirection.
The issue was that when the router is establishing or removing web redirection
at boot time, then this occurs before the init service is ready to process
rc notifications.  This would lead to dnsmasq not being running with the
proper redirection mode (or, at times, not even running at all).  Main symptom
was that VPN-based WAN was unable to resolve the pptp/l2tp server.

What this patch does is implement a force flag to start/stop_dnsmasq.  This flag
is used only in cases where we are restarting dnsmasq to accomodate a
potential redirection change, which means if it occurs during boot time,
it will still get properly restarted without having to go through the "rc"
process.