Retry failed iptables-restore in a few secs ; re-apply Samba-related rules (SMP only)
This resolves the collision occuring if another process tries to modify the
table while we are in the process of applying the firewall rules.  This mostly
occurs in SMP, where start_samba() tries to insert some rules at the same time
during boot.

Also we reapply the samba rules normally applied by start_samba() on ARM.