Build: Enable iproute2 for OpenVPN
OpenVPN configure defaults to iproute2 disabled and falls back to
using ifconfig.

This also fixes a bug where the configure script would detect the host's
path to ifconfig, which could be different from the target's path.
Because the path to `ip` is already explicitly set, this isn't an issue
for iproute.