rdnssd: Set ipv6_ra_conf to the value 'mset' when the managed RA flag is set
When ipv6_ra_conf is set to 'oset' the dhcp6c.conf file does not include an
IA_NA section which resulted in the dhcp6c not requesting an IPv6 WAN address.

This caused problems for at least one Singapore ISP and only resulted in
no IPv6 WAN address for Comcast users.