openvpn: Fix detection of whether a server/client instance was currently running.
This issue would cause the nvram state to get out of sync with reality as the router
could try to restart the instance, set the state to 0, then abort reporting that there
was a problem restarting the instance, leaving the state at 0 despite the instance being
actually running.  This could be seen by the webui trying to initialize and restart the
instance, and failing with an error.