Provide an empty WPAD URL for broken Windows/IE clients.
I considered making it default to http://router.asus.com/ which would have provided
out-of-the-box support for PAC (Asuswrt-Merlin supports it), but I feared that it
might have caused long delays for clients using the router DHCP but not its resolver.

This change in behaviour means that people using a customized WPAD option will need
to use a postconf script to modify the existing entry now instead of just appending
it to the existing options.