webui: When building the disk's partition list, also include unmounted partitions in the object's array, otherwise any indexed reference will be wrong since the webui's tree list takes into account all existing partitions, not just mounted ones.
Without this change, disks with unknown/unmounted partitions will generate an error when trying to edit sharing permissions.
For example, if partition 0 is unmounted (EFI, for example) and partition 1 contains the NTFS filesystem, selecting the first
permissions will make the webui try to access the second index entry - which wouldn't exist if we had skipped unmounted partitions
when building our newDisk.partition[] array.

The current fix is more a workaround, as the real issue lies in the actual design itself.

This issue was introduced with Asus' disk management rewrite in GPL 2769.