rc: Setup collection of prebuilt objects, and use them to setup the stage before compiling.
Asus is increasing the amount of code being moved to closed source objects.
This is unfortunately making things really difficult here, as it means I need
prebuilt objects for virtually every model that I want to compile for.

At this point we currently only have objects for the AC87U and AC3200.  We'll need GPL
releases for every other models to gather the missing objects.

I blame the FCC stricter rules, people trying to work around those rules,
and people buying competitor's hardware but running Asus's developped
(and sometimes proprietary) code on it.