syslog: Improvements to loglevel management.
By default, logmessage() always used Emergency (0), which means that anything logged by rc or httpd
would show up as an emergency entry when using an external syslog daemon.  This now defaults to
Notice (5), and is configurable on the System page.

Also made the loglevel of the local syslogd configurable through the webui (the nvram already existed).