firewall: Don't drop port 5060 traffic if the SIP passthrough is disabled, reverting a GPL 858 change.
Unfortunately, the NAT helpers are tied to the Passthrough settings, and there are people whose clients
do not work well with the NAT helper enabled.  Ideally we'd need the passthrough and NAT helpers to be
separate options.  Until then, the only fix for those users is to revert the behaviour to pre-858 GPL.

Ideally, the passthrough settings should be separate from the NAT helpers.