rc: Do not enforce b/g protection when wireless mode is set to "Auto" (fixes #986).
Technically, this checkbox sets it to either "auto" or "off".  So while having b/g protection
set to "auto" when the wireless mode (which can be N-only, Legacy-only or Auto)
is also set to "Auto" can make sense, this isn't intuitive for the end-user, as
the webui shows it as being either an "on" or "off" checkbox.