nano: Add Nano 2.6.3 and ncurses 6.0
For now this is only enabled for ARM models (due to their larger rootfs partition).
However since I was able to compile ncurses with a very limited subset of terminfo
definition, the resulting overhead is only 100 KB, so I might eventually enable it
also for the RT-N66U (which has a dynamic jffs partition) and RT-AC66U (32 MB rootfs).

ncurses is compiled with only a few builtin terminal definitions:
xterm, vt100, vt200 and linux.  This saves a lot of flash space, and
should be enough for most users.