firewall: Network Service rules would only be applied to Parental-controled clients if it was enabled.  Move rules into a separate chain, and jump to that chain from both the PControl (accept) and the FORWARD chain.
To keep changes to a minimum, the "chain" variable was reused.  A cleaner approach would be to
hardcode it in the fprint() calls instead, but that would generate too many code changes, which
would make it next to impossible to merge future changes made by Asus to that portion of the code.