openvpn: Change the "default" server cipher from OpenVPN's built-in default (currently BF-CBC with deprecated 64-bit blocks) to the more secure AES-128-CBC.
OpenVPN intends to change that built-in default in the future, in the meantime having
Asuswrt default to a stronger cipher for factory default values is the safest way to do
so without breaking existing configurations.  Existing configuration do not require
any change (unless one wants to better secure their tunnel).