Implement new firmware availability check.
This check is based on the same mechanism used by Asus's own check for
new firmware.  This include the recently added capacity of displaying
the changelog.  The code has been adapted to use our own version numbering
scheme (including alpha/beta extendno tags).  The update server is
currently not configured in this commit as further testing might be needed
before exposing the server to the public.

This check is opt-out - users can completely disable it on the
Tools -> Other Settings page.  The setting could not be put on
the Firmware Upgrade page itself as that page has no capacity of
saving any settings change (unless extensive changes were to be
made.)

New firmware notification will be displayed at the top of the webui.

The update server is https-based, and an invalid SSL certificate will
cause the check to fail (changed from Asus's implementation.)

This code will NOT allow a user to automatically download and flash new
firmwares - this is NOT planned, as there are too many security
considerations to be taken care of for such a feature.