openvpn: implement tls-crypt support
tls-crypt is new in OpenVPN 2.4.  It allows to encrypt the TLS control
channel traffic using a static key.  This is one step further what
tls-auth did, which was only validate the authenticity of a
control packet.  This can help obfuscate the OpenVPN traffic, and
increase security/privacy.

The same static key is used in either modes.