webui: display DDNS FQDN for Namecheap on the main networkmap page instead of just hostname
amendment to the show_ddns_status() function, to derive a more meaningful 'ddnsName' dependant on the DDNS provider, which will be displayed to the user on the 'Network Map' page - for example instead of displaying 'hostname', it will display 'hostname.domainname' in certain instances.

* Update index.asp

Further testing has determined that this patch is not required when using WWW.ASUS.COM as the DDNS provider. The NVRAM variable 'ddns_hostname_x' stores the FQDN i.e. 'testhost.asuscomm.com' despite 'Advanced_ASUSDDNS_Content.asp' only displaying 'testhost' in the 'Host Name' field.