httpd: create separate versions of closed source components, and populate prebuild/ at build time
With 7378, the RT-AC68U has a different version of web_hook.o.
It wouldn't be surprising if in the future more differences started
to appear with other models.

Note that since git stores each duplicate file as just one single file, there's
no point in using symlinks for identical files.  This will make it easier
to merge new GPL blobs in the future, by simply coping to the appropriate
model.