webui: hide internal switch ports that are returned by robocfg since 7743
Since all of our supported models are limited to 4 BCM ports, use that as
a limit.  We cannot use the lanports nvram because on the RT-AC88U the
port used to connect to the Realtek switch is also enumerated in that
nvram.

If a future model offers more than 4 BCM ports, we can adjust that limit
based on the router model.